// src/rhsm/models/Manifest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Manifest = {
    entitlementQuantity?: number;
    name?: string;
    type?: string;
    url?: string;
    uuid?: string;
    version?: string;
    simpleContentAccess?: string;
};