// src/rhsm/models/ManifestSummary.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Date } from './Date';
/**
 * details of a manifest
 */
export type ManifestSummary = {
    contentAccessMode?: string;
    createdBy?: string;
    createdDate?: Date;
    entitlementsAttachedQuantity?: number;
    lastModified?: Date;
    name?: string;
    type?: string;
    uuid?: string;
    version?: string;
};