// src/rhsm/models/poolsListMock.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { APIPageParam } from './APIPageParam';
import type { PoolDetail } from './PoolDetail';
export type poolsListMock = {
    body?: Array<PoolDetail>;
    pagination?: APIPageParam;
};